+++ lib/app-container/app-config.js
@@ -1,13 +1,14 @@
 'use strict';
-module.exports = function(mounter) {
+module.exports = function(configurator) {
   return {
     config: function(name) {
-      let config = {rootPath: mounter.rootPath};
-      config.page = mounter.page.create(name, config);
+      let config = {rootPath: configurator.rootPath};
+      let App = configurator.App;
+      config.page = new App(name, config);
       return {name: name, config: config};
     },
     configs: function(apps) {
-      apps = apps || mounter.appList;
+      apps = apps || configurator.appList;
       let self = this;
       // mounting multiple apps from same root dir, blank state
       return apps.map(function(name) {
