+++ lib/app-container/index.js
@@ -4,9 +4,6 @@ let extend = require('extend');
 // new appContainer(opts)
 module.exports = function(opts) {
   opts = opts || {debug: false};
-  if (!koaApp)
-    throw 'AppContainer constructor must take a koaApp as argument';
-
   let Mounter = require('./mounter');
   let Routes = require('./routes');
 
@@ -47,5 +44,6 @@ module.exports = function(opts) {
     Routes: Routes
   };
   appContainer.mount = new Mounter(appContainer);
+  console.log('appContainer', appContainer);
   return appContainer;
 }
