+++ lib/app-container/mounter.js
@@ -36,5 +36,5 @@ module.exports = function(appContainer) {
       }
       return appContainer;
     }
-  }
+  };
 }
