+++ lib/app-container/routes.js
@@ -1,10 +1,13 @@
 'use strict';
 module.exports = function (appContainer) {
   let koaApp = appContainer.koaApp;
+  if (!koaApp) throw 'AppContainer::Routes must take an AppContainer with a KoaApp configured';
+  console.log('koaApp', koaApp);
+  let Router = koaApp.Router;
   return {
     appContainer: appContainer,
     koaApp: koaApp,
-    routeCreator: new koaApp.Router(appContainer, koaApp.app),
+    routeCreator: new Router(appContainer, koaApp.app),
 
     // registry of routes to SPAs added
     routes: [],
