+++ lib/app/resolver/index.js
@@ -1,8 +1,8 @@
 'use strict';
 
 module.exports = function(name, config) {
-  let StateResolver = require('./state'),
-  let TemplateResolver = require('./template')
+  let StateResolver = require('./state');
+  let TemplateResolver = require('./template');
 
   let stateResolver = new StateResolver(name, config);
   let templateResolver = new TemplateResolver(name, config);
