+++ lib/index.js
@@ -15,7 +15,7 @@ module.exports = {
     },
     Resolver: require('./app/resolver'),
     utils: require('./utils')
-  }
+  },
   App: require('./app'),
   AppContainer: require('./app-container'),
   AppConfigurator: require('./app-container/configurator'),
